Xenoblade 3: added most voice event names
Unfortunately, everything is numeric and not that great to look at.

The vo_XXX_ files are divided up by character and have all audio
that isn't in a cutscene.  See gb/game/system.bdat for mapping.

All the other events are different types of cutscene, and their names
should be cross-referenced against the BDAT files in gb/evt.

This provides event names for all but about 40 of the cutscene and
environmental events.  Most of the cutscene ones seem to be duplicates
of other events, based on comparison with subtitle files.